requirements.txt
django~=4.1.7
drf-pydantic
psycopg2
openpyxl~=3.1.2
aiohttp~=3.8.4
asyncio~=3.4.3
djangorestframework~=3.14.0
pydantic~=1.10.6